Padronização de codigo

+++ src/controller/userController.js
@@ -1,54 +1,65 @@
-import Funcao from './functions'
+import Framework from './functions'
 import UserModel from '../model/user'
 import LogControl from './LogController'
 import AntigaControl from './SenhaAntigaController'
 
 //#region CRUD
+async function inserir(nome, email, senha) {
+    try {
+        let verificaCadastro
+        let historicoSenha
 
-async function inserir(nome, email, senha, senhaEncripta) {
-    let result
-    let historicoSenha
-    result = await login(email, senha)
-    if (result.message == "Usuário não cadastrado") {
-        historicoSenha = await AntigaControl.SenhaAntiga(email, senha)
-        if (historicoSenha.status == 400) {
-            return historicoSenha
+        verificaCadastro = await login(email, senha)
+        if (verificaCadastro.message == "Usuário não cadastrado") {
+            historicoSenha = await AntigaControl.SenhaAntiga(email, senha)
+            if (historicoSenha.status != 200)
+                return historicoSenha
+
+            senha = await Framework.ValidarDado("senha", senha)
+            if (senha.status != 200)
+                return senha
+
+            await UserModel.create({ email: email, nome: nome, senha: senha.result.senhacripta })
+            verificaCadastro = await login(email, senha.result.senha)
+            return verificaCadastro
+        } else {
+            return verificaCadastro
         }
-        nome = await Funcao.encripta(nome)
-        await UserModel.create({ email, nome, senha: senhaEncripta })
-        result = await login(email, senha)
-        return result
-    } else {
-        return result
+    } catch (e) {
+        return Framework.PadronizarRetorno("erro", 400, `Erro ao cadastrar usuário: ${e.message}`)
     }
 }
 
 async function login(email, senha) {
-    let oSenhaBd
-    let oBuscarUser
-    let token
-    oBuscarUser = await UserModel.findOne({ email })
-    if (oBuscarUser == "" || oBuscarUser == null) {
-        return Funcao.padraoErro("Usuário não cadastrado")
-    } else {
-        oSenhaBd = await Funcao.verificajwt(oBuscarUser.senha)
-        if (oSenhaBd == false) {
-            return Funcao.padraoErro("Senha Expirada")
-        }
-        if (oSenhaBd == senha) {
-            token = await Funcao.gerajwt(oBuscarUser._id)
+    try {
+        let oSenhaBd
+        let oBuscarUser
+        let token
+        oBuscarUser = await UserModel.findOne({ email })
+        if (oBuscarUser == "" || oBuscarUser == null)
+            return Framework.PadronizarRetorno("erro", 400, "Usuário não cadastrado")
+
+
+        oSenhaBd = await Framework.ManipularDado("desencripta", oBuscarUser.senha)
+        if (oSenhaBd.status != 200)
+            return oSenhaBd
+
+        if (oSenhaBd.result == senha) {
+            token = await Framework.ManipularToken("criar", oBuscarUser._id)
             await LogControl.deleta(email)
-            return Funcao.padraoSucesso({ token })
+            return Framework.PadronizarRetorno("sucesso", 200, { token: token.result })
         } else {
             await LogControl.InserirLog(email)
-            return Funcao.padraoErro("Senha Incorreta")
+            return Framework.PadronizarRetorno("erro", 400, "Senha Incorreta")
         }
+    } catch (e) {
+        return Framework.PadronizarRetorno("erro", 400, `Erro ao realizar login: ${e.message}`)
     }
 }
 
 async function AlterarSenha(email, senha, senhaAntiga) {
     let acesso
-    let senhaEncripta = await Funcao.geraSenha(senha)
+    let senhaEncripta = await Framework.geraSenha(senha)
     if (senhaAntiga == '' || senhaAntiga == null || senhaAntiga == undefined) {
         acesso = await login(email, senha)
         if (acesso.message == "Senha Expirada") {
@@ -58,7 +69,7 @@ async function AlterarSenha(email, senha, senhaAntiga) {
                 acesso = await login(email, senha)
                 return acesso
             }
-        } else if (acesso.message == "Senha Incorreta") {   
+        } else if (acesso.message == "Senha Incorreta") {
             let result = await UserModel.findOne({ email })
             if (result) {
                 await UserModel.findByIdAndUpdate({ _id: result._id }, { senha: senhaEncripta })
@@ -78,35 +89,45 @@ async function AlterarSenha(email, senha, senhaAntiga) {
             acesso = await login(email, senha)
             return acesso
         }
-        return Funcao.padraoErro("Ops, Algo deu errado!!")
+        return Framework.PadronizarRetorno("erro", 400, "Ops, Algo deu errado!!")
     }
 }
 
-async function excluirId(user) {
-    let email = await UserModel.findById(user)
-    let result = await UserModel.findByIdAndDelete(user)
-    await AntigaControl.DeletaAntiga(email.email)
-    return Funcao.padraoSucesso(result)
+async function excluirId(token) {
+    try {
+        let user = await Framework.ManipularToken("retornaId",token)
+
+        if(user.status != 200)        
+        return user
+
+        let email = await UserModel.findById(user.result)
+        let result = await UserModel.findByIdAndDelete(user.result)
+
+        await AntigaControl.DeletaAntiga(email.email)
+        return Framework.PadronizarRetorno("sucesso", 200, result)
+    } catch (e) {
+        return Framework.PadronizarRetorno("erro", 400, `Erro ao apagar usuário: ${e.message}`)
+    }
 }
 
 async function listarUm(user) {
     let formata = new Object();
     let user1 = await UserModel.findById(user)
     formata.id = user
-    formata.nome = await Funcao.verificajwt(user1.nome)
-    formata.email = await Funcao.verificajwt(user1.email)
-    return Funcao.padraoSucesso(formata)
+    formata.nome = await Framework.verificajwt(user1.nome)
+    formata.email = await Framework.verificajwt(user1.email)
+    return Framework.PadronizarRetorno("sucesso", 200, formata)
 }
 
 //#endregion
 
 //#region DESENVOLVIMENTO
-async function testeSenha(senha) {
-    senha = await Funcao.validaSenha(senha)
-    return senha
-}
+async function excluirUm(email,chave) {
+    chave = await Framework.ManipularToken("chaveSecreta",chave)
 
-async function excluirUm(email) {
+    if(chave.status !=200)
+        return Framework.PadronizarRetorno("erro", 400, "Acesso não permitido, contate o desenvolvedor para quaisquer duvidas.")
+
     let result = await UserModel.findOneAndDelete({ email })
     return result
 }
@@ -114,64 +135,49 @@ async function excluirUm(email) {
 async function listar() {
     let result = await UserModel.find({})
     if (result) {
-        return result
+        return Framework.PadronizarRetorno("sucesso", 200, result)
     } else {
-        return Funcao.padraoErro("Lista Vazia")
+        return Framework.PadronizarRetorno("erro", 400, "Lista Vazia")
     }
 }
-
 //#endregion
 
 async function Validador(token, nome, email, senha, senhaAntiga, rota) {
-    let user = await Funcao.verificajwt(token)
-    let testeEmail = Funcao.validaEmail(email)
-    let testeSenha = await Funcao.validaSenha(senha)
-    let logExp = await LogControl.ValidaAcesso(email)
-    let result
-    let atualizaToken
-    if (user == false && token != '') {
-        return Funcao.padraoErro("Credencial invalida!!")
-    } else {
-        atualizaToken = await Funcao.atualizajwt(token)
-    }
-    if (email != '' && testeEmail.status == 400) {
-        return testeEmail
-    }
-    if (senha != '' && testeSenha.status == 400) {
-        return testeSenha
-    }
-    if (email != '' && logExp.status == 400) {
-        return logExp
-    }
+    let validartoken = "sucesso"
+
+    //#region TRATAMENTO DE DADOS
+    token = await Framework.ManipularToken("atualizar", token)
+    if (token.status != 200)
+        validartoken = "erro"
+    //#endregion
+
+    //#region CHAMADA DE METODOS
     switch (rota) {
         case "inserir":
             await LogControl.deleta(email)
-            result = await inserir(nome, email, senha, testeSenha.result.senhacripta)
-            break;
+            return await inserir(nome, email, senha)
+
         case "login":
-            result = await login(email, senha)
-            break;
+            return await login(email, senha)
+
         case "AlterarSenha":
             await LogControl.deleta(email)
-            result = await AlterarSenha(email, senha, senhaAntiga)
-            break;
+            return await AlterarSenha(email, senha, senhaAntiga)
+
         case "deleteId":
             await LogControl.deleta(email)
             await AntigaControl.deletaAntiga(email)
-            result = await excluirId(user)
-            break
+            return await excluirId(token)
+
         case "listarUm":
-            await listarUm(user)
-            break;
+            return await listarUm(token)
+
         case "atualizaToken":
-            result = Funcao.padraoSucesso(atualizaToken)
-            break;
+            return Framework.PadronizarRetorno(validartoken, token.status, token.status == 200 ? token.result : token.message)
+
         default:
-            return padraoErro("[Api] Erro ao informar a rota.")
+            return Framework.PadronizarRetorno("erro", 400, "[Api] Erro ao informar a rota.")
     }
-    if(result==null||result==''|| result==undefined){
-        return Funcao.padraoErro("Erro na Rota "+rota+" não foi retornado dados. Contate o desenvolvedor.")
-    }
-    return result
+    //#endregion
 }
-module.exports = { Validador, excluirUm, listar, listarUm, testeSenha }
+module.exports = { Validador, excluirUm, listar}
